Neuron version 0.1.3.6: Implemented the login button functionality at the login page
Users can now log in from the login page with their email and password. I've encountered a minor problem when I couldn't log in the users, which had signed up using their google account, with their email and password. Therefore, in the process of FTGS, password is also collected and linked to the firebase user as an authentication provider in order to enable email-password login.